+++ view/frontend/web/js/view/cart/afterpay-minicart.js
@@ -7,14 +7,20 @@ define(
         "jquery",
         "Magento_Catalog/js/price-utils",
         'mage/url',
-        'Magento_Customer/js/customer-data'
+        'Magento_Customer/js/customer-data',
+        'Afterpay_Afterpay/js/view/express/button'
     ],
-    function ( $, priceUtils,mageUrl,customerData) {
+    function ($, priceUtils, mageUrl, customerData, expressBtn) {
         return function (config){
             function canDispayExpress(){
                 var minicart_subtotal = customerData.get('cart')().subtotalAmount;
 
-                if (!isAfterpayRestricted() && parseFloat(minicart_subtotal) >= parseFloat(config.afterpayConfig.minLimit) && parseFloat(minicart_subtotal) <= parseFloat(config.afterpayConfig.maxLimit) && (config.afterpayConfig.paymentActive)) {
+                if (!isAfterpayRestricted() &&
+                    parseFloat(minicart_subtotal) >= parseFloat(config.afterpayConfig.minLimit) &&
+                    parseFloat(minicart_subtotal) <= parseFloat(config.afterpayConfig.maxLimit) &&
+                    (config.afterpayConfig.paymentActive) &&
+                    expressBtn.canDisplayOnMinicart()
+                ) {
                     $('#afterpay-minicart-express-button').show();
                 } else {
                     $('#afterpay-minicart-express-button').hide();