+++ view/frontend/web/js/view/payment/method-renderer/afterpaypayovertime.js
@@ -2,7 +2,7 @@
  * Magento 2 extensions for Afterpay Payment
  *
  * @author Afterpay
- * @copyright 2016-2020 Afterpay https://www.afterpay.com
+ * @copyright 2016-2021 Afterpay https://www.afterpay.com
  */
 /*browser:true*/
 /*global define*/
@@ -196,7 +196,7 @@ define(
                             email = document.getElementById("customer-email").value;
                         }
 
-                        data = data + '&email=' + email;
+                        data = data + '&email=' + encodeURIComponent(email);
 
 
                         $.ajax({