+++ view/frontend/web/js/view/payment/method-renderer/afterpayredirect.js
@@ -2,7 +2,7 @@
  * Magento 2 extensions for Afterpay Payment
  *
  * @author Afterpay
- * @copyright 2016-2020 Afterpay https://www.afterpay.com
+ * @copyright 2016-2021 Afterpay https://www.afterpay.com
  */
 /*browser:true*/
 /*global define*/